Change the installation logic to target decorview instead of android.R.id.content, since some apps use that view for fragment transactions.
