+++ hyperion-sample/src/main/java/com/willowtreeapps/hyperion/sample/MainActivity.java
@@ -2,47 +2,17 @@ package com.willowtreeapps.hyperion.sample;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentStatePagerAdapter;
-import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
 
 public class MainActivity extends AppCompatActivity {
 
-    private ViewPager pager;
-    private MainAdapter adapter;
-
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-        pager = findViewById(R.id.pager);
-        pager.setAdapter(adapter = new MainAdapter(getSupportFragmentManager()));
-    }
-
-    private static final class MainAdapter extends FragmentStatePagerAdapter {
-
-        public MainAdapter(FragmentManager fm) {
-            super(fm);
-        }
-
-        @Override
-        public Fragment getItem(int position) {
-            switch (position) {
-                case 0:
-                    return new MainFragment();
-                case 1:
-                    return new GestureFragment();
-                case 2:
-                    return new VersionFragment();
-            }
-            return null;
-        }
-
-        @Override
-        public int getCount() {
-            return 3;
+        if (savedInstanceState == null) {
+            getSupportFragmentManager().beginTransaction()
+                    .add(android.R.id.content, new PagerFragment())
+                    .commit();
         }
     }
 }
