+++ hyperion-sample/src/main/res/layout/fragment_pager.xml
